+++ lib/flutter_brand_palettes.dart
@@ -2,5 +2,4 @@ export 'brands.dart';
 export 'colors.dart';
 export 'src/palette.dart';
 export 'src/palette_rgb.dart';
-export 'src/swatch.dart';
 export 'swatches.dart';