+++ lib/swatches.dart
@@ -1,1 +1,3 @@
 export 'src/material_design/swatches/primaries.dart';
+export 'src/swatch.dart';
+export 'src/swatch_base.dart';